refactor(frontend/api): 문제 응답 정규화 로직 공통화
